refactor(components): rename components
