+++ src/components/WipHistogram.vue
